Test sorting persons. Selectors and methods on persons page.
